CI: Remove aws lambda packages from oss
