+++ tests/ci/build_check.py
@@ -12,7 +12,6 @@ import docker_images_helper
 from ci_config import CI
 from env_helper import REPO_COPY, S3_BUILDS_BUCKET, TEMP_PATH
 from git_helper import Git
-from lambda_shared_package.lambda_shared.pr import Labels
 from pr_info import PRInfo
 from report import FAILURE, SUCCESS, JobReport, StatusType
 from stopwatch import Stopwatch
@@ -108,7 +107,9 @@ def build_clickhouse(
 
 
 def is_release_pr(pr_info: PRInfo) -> bool:
-    return Labels.RELEASE in pr_info.labels or Labels.RELEASE_LTS in pr_info.labels
+    return (
+        CI.Labels.RELEASE in pr_info.labels or CI.Labels.RELEASE_LTS in pr_info.labels
+    )
 
 
 def get_release_or_pr(pr_info: PRInfo, version: ClickHouseVersion) -> Tuple[str, str]:
