--- tests/ci/cancel_and_rerun_workflow_lambda/requirements.txt
@@ -1,1 +1,0 @@
-../lambda_shared_package[token]
