--- tests/ci/ci_runners_metrics_lambda/requirements.txt
@@ -1,2 +1,0 @@
-../lambda_shared_package
-../lambda_shared_package[token]
