--- tests/ci/clean_lost_instances_lambda/lambda_shared
@@ -1,1 +1,0 @@
-../lambda_shared_package/lambda_shared
