--- tests/ci/clean_lost_instances_lambda/requirements.txt
@@ -1,2 +1,0 @@
-../lambda_shared_package
-../lambda_shared_package[token]
