--- tests/ci/lambda_shared_package/.gitignore
@@ -1,2 +1,0 @@
-build
-*.egg-info
