--- tests/ci/lambda_shared_package/pyproject.toml
@@ -1,24 +1,0 @@
-[build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "lambda_shared"
-version = "0.0.1"
-dependencies = [
-  "requests",
-  "urllib3 < 2"
-]
-
-[project.optional-dependencies]
-token = [
-  "PyJWT",
-  "cryptography",
-]
-dev = [
-  "boto3",
-  "lambda_shared[token]",
-]
-
-[tool.distutils.bdist_wheel]
-universal = true
