--- tests/ci/lambda_shared_package/setup.cfg
@@ -1,8 +1,0 @@
-### This file exists for clear builds in docker ###
-#   without it the `build` directory wouldn't be  #
-# updated on the fly and will require manual clean #
-[build]
-build_base = /tmp/lambda_shared
-
-[egg_info]
-egg_base = /tmp/
