--- tests/ci/slack_bot_ci_lambda/requirements.txt
@@ -1,1 +1,0 @@
-../lambda_shared_package
