--- tests/ci/team_keys_lambda/requirements.txt
@@ -1,1 +1,0 @@
-../lambda_shared_package
