--- tests/ci/terminate_runner_lambda/requirements.txt
@@ -1,1 +1,0 @@
-../lambda_shared_package[token]
