--- tests/ci/workflow_approve_rerun_lambda/requirements.txt
@@ -1,1 +1,0 @@
-../lambda_shared_package[token]
